Updates from Oct 5
I have added some changes to the notebooks from the comments in Word Document sent on October 5th, 2023. All Notebooks run as long as the "wsc" environment is activated. The FGF_Manual_Peak_selector notebook should be removed as it does not generate any figure or information to the paper.